Use shared component
