+++ src/panels/config/automation/ha-automation-trace.ts
@@ -1,26 +1,18 @@
 import "@home-assistant/webawesome/dist/components/divider/divider";
 import {
-  mdiAlertCircleOutline,
-  mdiCheckCircleOutline,
-  mdiChevronDown,
   mdiDotsVertical,
   mdiDownload,
-  mdiHelpCircleOutline,
   mdiInformationOutline,
   mdiPencil,
-  mdiProgressClock,
-  mdiProgressWrench,
   mdiRayEndArrow,
   mdiRayStartArrow,
   mdiRefresh,
-  mdiStopCircleOutline,
 } from "@mdi/js";
 import type { CSSResultGroup, TemplateResult, PropertyValues } from "lit";
 import { css, html, LitElement, nothing } from "lit";
 import { customElement, property, query, state } from "lit/decorators";
 import { classMap } from "lit/directives/class-map";
 import { isComponentLoaded } from "../../../common/config/is_component_loaded";
-import { formatDateTimeWithSeconds } from "../../../common/datetime/format_date_time";
 import { fireEvent } from "../../../common/dom/fire_event";
 import { navigate } from "../../../common/navigate";
 import { computeRTL } from "../../../common/util/compute_rtl";
@@ -53,9 +45,7 @@ import "../../../layouts/hass-subpage";
 import { haStyle } from "../../../resources/styles";
 import type { HomeAssistant, Route } from "../../../types";
 import { fileDownload } from "../../../util/file_download";
-import "../../../components/ha-generic-picker";
-import type { PickerComboBoxItem } from "../../../components/ha-picker-combo-box";
-import type { HaGenericPicker } from "../../../components/ha-generic-picker";
+import "../../../components/ha-trace-picker";
 
 const TABS = ["details", "timeline", "logbook", "automation_config"] as const;
 
@@ -86,8 +76,6 @@ export class HaAutomationTrace extends LitElement {
   @state() private _logbookEntries?: LogbookEntry[];
 
   @state() private _view: (typeof TABS)[number] | "blueprint" = "details";
-
-  @query("ha-generic-picker") private tracePicker?: HaGenericPicker;
 
   @query("hat-script-graph") private _graph?: HatScriptGraph;
 
@@ -191,32 +179,12 @@ export class HaAutomationTrace extends LitElement {
                   this._runId}
                   @click=${this._pickOlderTrace}
                 ></ha-icon-button>
-                <ha-generic-picker
-                  name="trace"
+                <ha-trace-picker
                   .hass=${this.hass}
-                  .label=${this.hass.localize(
-                    "ui.panel.config.automation.trace.select_trace"
-                  )}
+                  .traces=${this._traces}
                   .value=${this._runId}
-                  .getItems=${this._getTraces}
-                  required
                   @value-changed=${this._pickTrace}
-                >
-                  <ha-button
-                    slot="field"
-                    appearance="filled"
-                    variant="neutral"
-                    size="small"
-                    @click=${this._openPicker}
-                  >
-                    ${this._renderTracePickerValue(this._runId!)}
-                    <ha-svg-icon
-                      slot="end"
-                      .path=${mdiChevronDown}
-                    ></ha-svg-icon>
-                  </ha-button>
-                </ha-generic-picker>
-
+                ></ha-trace-picker>
                 <ha-icon-button
                   .label=${this.hass!.localize(
                     "ui.panel.config.automation.trace.newer_trace"
@@ -343,118 +311,6 @@ export class HaAutomationTrace extends LitElement {
       </hass-subpage>
     `;
   }
-
-  private _openPicker(ev: Event) {
-    ev.stopPropagation();
-    this.tracePicker?.open();
-  }
-
-  private _getTraces = (): PickerComboBoxItem[] =>
-    this._traces?.map((trace) => {
-      const renderRuntime = () =>
-        (
-          (new Date(trace.timestamp.finish!).getTime() -
-            new Date(trace.timestamp.start).getTime()) /
-          1000
-        ).toFixed(2);
-
-      const item: PickerComboBoxItem = {
-        id: trace.run_id,
-        primary: formatDateTimeWithSeconds(
-          new Date(trace.timestamp.start),
-          this.hass.locale,
-          this.hass.config
-        ),
-      };
-      if (trace.state === "running") {
-        item.secondary = this.hass.localize(
-          "ui.panel.config.automation.trace.picker.still_running"
-        );
-        item.icon_path = mdiProgressClock;
-      } else if (trace.state === "debugged") {
-        item.secondary = this.hass.localize(
-          "ui.panel.config.automation.trace.picker.debugged"
-        );
-        item.icon_path = mdiProgressWrench;
-      } else if (trace.script_execution === "finished") {
-        item.secondary = this.hass.localize(
-          "ui.panel.config.automation.trace.picker.finished",
-          {
-            executiontime: renderRuntime(),
-          }
-        );
-        item.icon_path = mdiCheckCircleOutline;
-      } else if (trace.script_execution === "aborted") {
-        item.secondary = this.hass.localize(
-          "ui.panel.config.automation.trace.picker.aborted",
-          {
-            executiontime: renderRuntime(),
-          }
-        );
-        item.icon_path = mdiAlertCircleOutline;
-      } else if (trace.script_execution === "cancelled") {
-        item.secondary = this.hass.localize(
-          "ui.panel.config.automation.trace.picker.cancelled",
-          {
-            executiontime: renderRuntime(),
-          }
-        );
-        item.icon_path = mdiAlertCircleOutline;
-      } else {
-        let message:
-          | "stopped_failed_conditions"
-          | "stopped_failed_single"
-          | "stopped_failed_max_runs"
-          | "stopped_error"
-          | "stopped_unknown_reason";
-        let error: string | undefined;
-        let icon: string;
-
-        switch (trace.script_execution) {
-          case "failed_conditions":
-            message = "stopped_failed_conditions";
-            icon = mdiStopCircleOutline;
-            break;
-          case "failed_single":
-            message = "stopped_failed_single";
-            icon = mdiStopCircleOutline;
-            break;
-          case "failed_max_runs":
-            message = "stopped_failed_max_runs";
-            icon = mdiStopCircleOutline;
-            break;
-          case "error":
-            message = "stopped_error";
-            error = trace.error!;
-            icon = mdiAlertCircleOutline;
-            break;
-          default:
-            message = "stopped_unknown_reason";
-            icon = mdiHelpCircleOutline;
-        }
-
-        item.secondary = this.hass.localize(
-          `ui.panel.config.automation.trace.picker.${message}`,
-          {
-            error,
-            executiontime: renderRuntime(),
-          }
-        );
-        item.icon_path = icon;
-      }
-      return item;
-    }) ?? [];
-
-  private _renderTracePickerValue = (runId: string) => {
-    const trace = this._traces?.find((t) => t.run_id === runId);
-    return html`${trace
-      ? formatDateTimeWithSeconds(
-          new Date(trace.timestamp.start),
-          this.hass.locale,
-          this.hass.config
-        )
-      : runId}`;
-  };
 
   protected firstUpdated(changedProps: PropertyValues<this>) {
     super.firstUpdated(changedProps);
@@ -753,12 +609,9 @@ export class HaAutomationTrace extends LitElement {
         ha-trace-logbook {
           direction: var(--direction);
         }
-        ha-generic-picker {
+        ha-trace-picker {
           flex-grow: 1;
           max-width: 500px;
-        }
-        ha-generic-picker > ha-button {
-          width: 100%;
         }
       `,
     ];
